feat(frontend): Clean Blog card
