+++ src/css/homepage.css
@@ -5,18 +5,104 @@
 
 @import url("https://fonts.googleapis.com/css?family=Roboto:300,400,500,700&display=swap");
 
-/* Title, StartResource */
-#section-title-startResource-id {
+.bodyLight {
     font-family: Roboto;
-    font-size: 51.4px;
-    font-weight: 600;
     font-stretch: normal;
     font-style: normal;
-    line-height: normal;
-    letter-spacing: normal;
+    text-align: left;
+    font-weight: normal;
+    color: #333;
+    font-size: 14px;
+    line-height: 1.45;
+    letter-spacing: 0.66px;
 }
 
-/* Blue box from title StartResource */
+.bodyDark {
+    font-family: Roboto;
+    font-stretch: normal;
+    font-style: normal;
+    text-align: left;
+    font-weight: normal;
+    color: #f2f2f2;
+    font-size: 14px;
+    line-height: 1.45;
+    letter-spacing: 0.66px;
+}
+
+.overlineLight {
+    font-family: Roboto;
+    font-stretch: normal;
+    font-style: normal;
+    text-align: left;
+    font-weight: normal;
+    color: #333;
+    font-size: 18.4px;
+    line-height: 1.71;
+    letter-spacing: 0.66px;
+}
+
+.overlineDark {
+    font-family: Roboto;
+    font-stretch: normal;
+    font-style: normal;
+    text-align: left;
+    font-weight: normal;
+    color: #f2f2f2;
+    font-size: 18.4px;
+    line-height: 1.71;
+    letter-spacing: 0.66px;
+}
+
+.titleLight {
+    font-family: Roboto;
+    font-stretch: normal;
+    font-style: normal;
+    text-align: left;
+    font-weight: bold;
+    color: #000;
+    font-size: 21px;
+    line-height: 1.09;
+    letter-spacing: 0.13px;
+    margin-top: 2px;
+}
+
+.titleDark {
+    font-family: Roboto;
+    font-stretch: normal;
+    font-style: normal;
+    text-align: left;
+    font-weight: bold;
+    color: #fff;
+    font-size: 21px;
+    line-height: 1.09;
+    letter-spacing: 0.13px;
+    margin-top: 2px;
+}
+
+.linkLight {
+    color: #000;
+}
+
+.linkDark {
+    color: #fff;
+}
+
+#link-id-customCard {
+    font-family: Roboto;
+    font-size: 18.4px;
+    font-weight: 500;
+    line-height: 1.43;
+    letter-spacing: 0.13px;
+    cursor: pointer;
+}
+
+.blueBoxTitle {
+    width: 3px;
+    background-color: #2d9cdb;
+    height: 40px;
+    margin: 0 16px 0 0;
+}
+
 .box-title-startResource {
     width: 9px;
     height: 60px;
@@ -25,14 +111,18 @@
     margin-right: 16px;
 }
 
-/* customCard dimentions */
 .cardroot {
     width: 360px;
     height: 240px;
     position: relative;
 }
 
-/* This is for Link berfore the efect, customCard */
+.cardroot:hover #box-link-id-customCard::before {
+    visibility: visible;
+    transform: scaleX(1);
+    cursor: pointer;
+}
+
 #box-link-id-customCard::before {
     content: "";
     position: absolute;
@@ -45,133 +135,47 @@
     transition: .25s linear;
 }
 
-/* This for linkEffect, customCard */
-.cardroot:hover #box-link-id-customCard::before {
-    visibility: visible;
-    transform: scaleX(1);
-    cursor: pointer;
-}
-
-/* Link style customCard */
-#link-id-customCard {
-    font-family: Roboto;
-    font-size: 18.4px;
-    font-weight: 500;
-    line-height: 1.43;
-    letter-spacing: 0.13px;
-    cursor: pointer;
-}
-
-/* This changes customCard header background */
-.cardroot:hover .header-transition {
-    background-color: #2d9cdb;
-}
-
-/* This changes customCard header title */
-.cardroot:hover #header-title {
-    color: #f2f2f2;
-}
-
-/* Header colors transictions customCard */
-.header-transition {
-    height: 56px;
-    transition: all .25s ease-in-out;
-}
-
-/* size customCard images */
-.card-img {
-    width: 42;
-    height: 42;
-}
-
-/* Container image customCard */
-.card-img-container {
-    width: 84.2px;
-    height: 84.2px;
-    flex-grow: 0;
-    margin: 13px 16.3px 14.4px calc(100% - 101px);
-    margin-right: 16.8;
-    padding: 23px 13.7px 23.4px 14.5px;
-    border: solid 2px #333;
-    background-color: #fff;
-    border-radius: 50%;
-    display: flex;
-    justify-content: center;
-    align-items: center;
-    position: absolute;
-}
-
-/* Header title, customCard */
-#header-title {
-    font-family: Roboto;
-    font-size: 18.4px;
-    line-height: 1.71;
-    letter-spacing: 0.66px;
-}
-
-#imgCarousel-id {
-    min-height: 175px;
-}
-
-@media screen and (max-width: 720px) {
-
-    .card-img {
-        width: 25;
-        height: 25;
-    }
-
-    .card-img-container {
-        width: 54.2px;
-        height: 54.2px;
-        margin: 10px 16.3px 14.4px calc(100% - 70.6px);
-        /* padding: 11px 7.6px 11.1px; */
-        padding: 2px 6px;
-        transition: all .25s ease-in-out;
-    }
-
-    #section-title-startResource-id {
-        font-size: 35px;
-        font-weight: 500;
-    }
-
-    .header-transition {
-        height: 40px;
-    }
-}
-
-@media screen and (max-width: 600px) {
-    .card-img {
-        width: 42;
-        height: 42;
-    }
-
-    /* Container image customCard */
-    .card-img-container {
-        width: 84.2px;
-        height: 84.2px;
-        flex-grow: 0;
-        margin: 13px 16.3px 14.4px calc(100% - 101px);
-        padding: 23px 13.7px 23.4px 14.5px;
-        margin-right: 16.8;
-        transition: all .25s ease-in-out;
-    }
-
-    .header-transition {
-        height: 56px;
-    }
-}
 
 @media screen and (max-width: 400px) {
-
-    #section-title-startResource-id {
-        font-size: 18px;
-        font-weight: 500;
+    .bodyLight {
+        font-size: 12px;
+        line-height: 1.33;
+        letter-spacing: 0.4px;
     }
 
-    .box-title-startResource {
-        width: 3.3px;
-        height: 22.3px;
-        margin-right: 8px;
+    .bodyDark {
+        font-size: 12px;
+        line-height: 1.33;
+        letter-spacing: 0.4px;
+    }
+
+    .overlineLight  {
+        font-size: 12px;
+        line-height: 1.16;
+        letter-spacing: 0.29px;
+    }
+
+    .overlineDark  {
+        font-size: 12px;
+        line-height: 1.16;
+        letter-spacing: 0.29px;
+    }
+
+    .titleLight {
+        font-size: 14px;
+        line-height: 0.99;
+        letter-spacing: 0.06px;
+    }
+
+    .titleDark {
+        font-size: 14px;
+        line-height: 0.99;
+        letter-spacing: 0.06px;
+    }
+
+    .blueBoxTitle {
+        height: 20px;
+        margin: 0 5px 0 0;
     }
 
     #link-id-customCard {
@@ -181,27 +185,10 @@
         letter-spacing: 0.07px;
     }
 
-    #header-title {
-        font-size: 11.1px;
-        line-height: 1.26;
-        letter-spacing: 0.29px;
+    .box-title-startResource {
+        width: 3.3px;
+        height: 22.3px;
+        margin-right: 8px;
     }
 
-    .card-img {
-        width: 21;
-        height: 18;
-    }
-
-    .card-img-container {
-        width: 40.2px;
-        height: 40.2px;
-        flex-grow: 0;
-        margin: 4px 16.3px 14.4px calc(100% - 54.6px);
-        padding: 5px 2px;
-        transition: all .25s ease-in-out;
-    }
-
-    .header-transition {
-        height: 24px;
-    }
 }
