+++ src/pages/heroSection/ImagesCarousel.js
@@ -34,7 +34,7 @@ class Carousel {
   }
 }
 
-export const ImagesCarousel = ({ background, index, displayImg }) => {
+const ImagesCarousel = ({ background, index, displayImg }) => {
   const { fade, numbertext } = useStyles();
 
   useEffect(() => {
@@ -44,9 +44,11 @@ export const ImagesCarousel = ({ background, index, displayImg }) => {
   }, []);
 
   return (
-    <Box key={background} className={`slideClass ${fade}`} display={displayImg} >
+    <Box key={background} className={`slideClass ${fade}`} display={displayImg}>
       <Box className={numbertext}>{index + 1} / 3</Box>
       <img id="imgCarousel-id" src={useBaseUrl(background)} width="100%" />
     </Box>
   );
 };
+
+export default ImagesCarousel;
